+++ core/imports/contacts.go
@@ -52,19 +52,19 @@ func ImportBatch(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets,
 	// gather up contacts and modifiers
 	mcs := make([]*models.Contact, 0, len(imports))
 	contacts := make([]*flows.Contact, 0, len(imports))
-	mods := make(map[flows.ContactUUID][]flows.Modifier, len(imports))
+	mods := make(map[models.ContactID][]flows.Modifier, len(imports))
 	for _, imp := range imports {
 		// ignore errored imports which couldn't get/create a contact
 		if imp.contact != nil {
 			mcs = append(mcs, imp.contact)
 			contacts = append(contacts, imp.flowContact)
-			mods[imp.flowContact.UUID()] = imp.mods
+			mods[imp.contact.ID()] = imp.mods
 			importsByContact[imp.flowContact] = imp
 		}
 	}
 
 	// and apply in bulk
-	eventsByContact, err := runner.BulkModify(ctx, rt, oa, userID, mcs, contacts, mods, models.ViaImport)
+	eventsByContact, err := runner.ModifyWithoutLock(ctx, rt, oa, userID, mcs, contacts, mods, models.ViaImport)
 	if err != nil {
 		return fmt.Errorf("error applying modifiers: %w", err)
 	}