+++ core/runner/scene.go
@@ -1,9 +1,12 @@
 package runner
 
 import (
+	"cmp"
 	"context"
 	"fmt"
 	"log/slog"
+	"maps"
+	"slices"
 	"time"
 
 	"github.com/nyaruka/goflow/flows"
@@ -11,7 +14,12 @@ import (
 	"github.com/nyaruka/goflow/flows/modifiers"
 	"github.com/nyaruka/mailroom/core/goflow"
 	"github.com/nyaruka/mailroom/core/models"
+	"github.com/nyaruka/mailroom/core/runner/clocks"
 	"github.com/nyaruka/mailroom/runtime"
+)
+
+const (
+	commitTimeout = time.Minute
 )
 
 // Scene represents the context that events are occurring in
@@ -37,6 +45,7 @@ type Scene struct {
 
 	preCommits    map[PreCommitHook][]any
 	postCommits   map[PostCommitHook][]any
+	rawEvents     []flows.Event
 	persistEvents []*models.Event
 
 	// can be overridden by tests
@@ -51,6 +60,7 @@ func NewScene(dbContact *models.Contact, contact *flows.Contact) *Scene {
 
 		preCommits:  make(map[PreCommitHook][]any),
 		postCommits: make(map[PostCommitHook][]any),
+		rawEvents:   make([]flows.Event, 0, 5),
 
 		Engine: goflow.Engine,
 	}
@@ -75,10 +85,13 @@ func (s *Scene) SprintUUID() flows.SprintUUID {
 	return s.Sprint.UUID()
 }
 
+// Events returns all events added to this scene (includes non-persisted events)
+func (s *Scene) Events() []flows.Event { return s.rawEvents }
+
 func (s *Scene) AddEvent(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, e flows.Event, userID models.UserID, via models.Via) error {
-	handler, found := eventHandlers[e.Type()]
-	if !found {
-		return fmt.Errorf("unable to find handler for event type: %s", e.Type())
+	handler := eventHandlers[e.Type()]
+	if handler == nil {
+		panic(fmt.Sprintf("no handler for event type: %s", e.Type()))
 	}
 
 	if err := handler(ctx, rt, oa, s, e, userID); err != nil {
@@ -93,12 +106,18 @@ func (s *Scene) AddEvent(ctx context.Context, rt *runtime.Runtime, oa *models.Or
 
 	e.SetUser(user.Reference(), string(via))
 
-	if models.PersistEvent(e) {
-		s.persistEvents = append(s.persistEvents, &models.Event{
-			Event:       e,
-			OrgID:       oa.OrgID(),
-			ContactUUID: s.ContactUUID(),
-		})
+	switch e.(type) {
+	case *ContactInterruptedEvent, *SprintEndedEvent: // our pseudo events aren't real...
+	default:
+		s.rawEvents = append(s.rawEvents, e)
+
+		if models.PersistEvent(e) {
+			s.persistEvents = append(s.persistEvents, &models.Event{
+				Event:       e,
+				OrgID:       oa.OrgID(),
+				ContactUUID: s.ContactUUID(),
+			})
+		}
 	}
 
 	return nil
@@ -128,7 +147,7 @@ func (s *Scene) addSprint(ctx context.Context, rt *runtime.Runtime, oa *models.O
 	return nil
 }
 
-func (s *Scene) Interrupt(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, status flows.SessionStatus) error {
+func (s *Scene) InterruptWaiting(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, status flows.SessionStatus) error {
 	return addInterruptEvents(ctx, rt, oa, []*Scene{s}, status)
 }
 
@@ -166,7 +185,7 @@ func (s *Scene) ResumeSession(ctx context.Context, rt *runtime.Runtime, oa *mode
 			// if flow doesn't exist, we can't resume, so fail the session
 			slog.Debug("unable to find flow for resume", "contact", s.ContactUUID(), "session", session.UUID, "flow", session.CurrentFlowUUID)
 
-			if err := s.Interrupt(ctx, rt, oa, flows.SessionStatusFailed); err != nil {
+			if err := s.InterruptWaiting(ctx, rt, oa, flows.SessionStatusFailed); err != nil {
 				return fmt.Errorf("error adding interrupt events for unresumable session %s: %w", session.UUID, err)
 			}
 
@@ -200,7 +219,7 @@ func (s *Scene) ResumeSession(ctx context.Context, rt *runtime.Runtime, oa *mode
 	return nil
 }
 
-func (s *Scene) ApplyModifier(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, mod flows.Modifier, userID models.UserID, via models.Via) ([]flows.Event, error) {
+func (s *Scene) ApplyModifier(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, mod flows.Modifier, userID models.UserID, via models.Via) error {
 	env := flows.NewAssetsEnvironment(oa.Env(), oa.SessionAssets())
 	eng := goflow.Engine(rt)
 
@@ -208,7 +227,7 @@ func (s *Scene) ApplyModifier(ctx context.Context, rt *runtime.Runtime, oa *mode
 	evtLog := func(e flows.Event) { evts = append(evts, e) }
 
 	if _, err := modifiers.Apply(ctx, eng, env, oa.SessionAssets(), s.Contact, mod, evtLog); err != nil {
-		return nil, fmt.Errorf("error applying %s modifier to contact %s: %w", mod.Type(), s.Contact.UUID(), err)
+		return fmt.Errorf("error applying %s modifier to contact %s: %w", mod.Type(), s.Contact.UUID(), err)
 	}
 
 	for _, e := range evts {
@@ -220,11 +239,11 @@ func (s *Scene) ApplyModifier(ctx context.Context, rt *runtime.Runtime, oa *mode
 		}
 
 		if err := s.AddEvent(ctx, rt, oa, e, creditUserID, via); err != nil {
-			return nil, fmt.Errorf("error adding modifier events for contact %s: %w", s.Contact.UUID(), err)
+			return fmt.Errorf("error adding modifier events for contact %s: %w", s.Contact.UUID(), err)
 		}
 	}
 
-	return evts, nil
+	return nil
 }
 
 // AttachPreCommitHook adds an item to be handled by the given pre commit hook
@@ -264,6 +283,61 @@ func CreateScenes(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets
 	}
 
 	return scenes, nil
+}
+
+// LockAndLoad tries to lock and load scenes for the given contact ids, returning any ids that could not be locked.
+// The caller is responsible for unlocking the scenes by calling the returned unlock function.
+func LockAndLoad(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, ids []models.ContactID, includeTickets map[models.ContactID][]*models.Ticket, timeout time.Duration) ([]*Scene, []models.ContactID, func(), error) {
+	allScenes := make([]*Scene, 0, len(ids))
+	allLocks := make(map[models.ContactID]string, len(ids))
+	allGood := false
+
+	unlockAll := func() {
+		if err := clocks.Unlock(context.Background(), rt, oa, allLocks); err != nil {
+			slog.Error("error unlocking contacts", "error", err, "contacts", maps.Keys(allLocks))
+		}
+	}
+
+	defer func() {
+		if !allGood {
+			unlockAll()
+		}
+	}()
+
+	remaining := ids
+	start := time.Now()
+
+	for len(remaining) > 0 && time.Since(start) < timeout {
+		if ctx.Err() != nil {
+			return nil, nil, nil, ctx.Err()
+		}
+
+		// try to get locks for these contacts, waiting for up to a second for each contact
+		locks, skipped, err := clocks.TryToLock(ctx, rt, oa, remaining, time.Second)
+		if err != nil {
+			return nil, nil, nil, err
+		}
+		maps.Copy(allLocks, locks)
+
+		locked := slices.Collect(maps.Keys(locks))
+
+		// create scenes for the locked contacts
+		scenes, err := CreateScenes(ctx, rt, oa, locked, includeTickets)
+		if err != nil {
+			return nil, nil, nil, fmt.Errorf("error creating scenes after locking: %w", err)
+		}
+
+		allScenes = append(allScenes, scenes...)
+
+		remaining = skipped // skipped are now our remaining
+	}
+
+	// for test determinism
+	slices.SortFunc(allScenes, func(a, b *Scene) int { return cmp.Compare(a.Contact.ID(), b.Contact.ID()) })
+
+	allGood = true // to prevent unlock in defer
+
+	return allScenes, remaining, unlockAll, nil
 }
 
 // BulkCommit commits the passed in scenes in a single transaction. If that fails, it retries committing each scene one at a time.