+++ web/contact/create.go
@@ -54,8 +54,8 @@ func handleCreate(ctx context.Context, rt *runtime.Runtime, r *createRequest) (a
 		return nil, 0, err
 	}
 
-	modifiers := map[flows.ContactUUID][]flows.Modifier{contact.UUID(): c.Mods}
-	_, err = runner.BulkModify(ctx, rt, oa, r.UserID, []*models.Contact{mc}, []*flows.Contact{contact}, modifiers, r.Via)
+	modifiers := map[models.ContactID][]flows.Modifier{mc.ID(): c.Mods}
+	_, err = runner.ModifyWithoutLock(ctx, rt, oa, r.UserID, []*models.Contact{mc}, []*flows.Contact{contact}, modifiers, r.Via)
 	if err != nil {
 		return nil, 0, fmt.Errorf("error modifying new contact: %w", err)
 	}